requirement.txt
opencv-python
pillow
numpy
tensorflow
SpeechRecognition
pyaudio
nltk
streamlit
spacy
keras
